+++ shuup/core/excs.py
@@ -55,6 +55,8 @@ class ProductNotOrderableProblem(Problem):
 class ProductNotVisibleProblem(Problem):
     pass
 
+class InvalidOrderStatusError(Problem):
+    pass
 
 class ImpossibleProductModeException(ValueError):
     def __init__(self, message, code=None):